ai-elderly-medicare-system/streamlit_app/app.py
try:
    import streamlit as st
    st_installed = True
except ImportError:
    st_installed = False
    print("Streamlit is not installed. Please run: pip install streamlit")

try:
    import streamlit_authenticator as stauth
    stauth_installed = True
except ImportError:
    stauth_installed = False
    print("Streamlit Authenticator is not installed. Please run: pip install streamlit-authenticator")

try:
    import yaml
    from yaml.loader import SafeLoader
    yaml_installed = True
except ImportError:
    yaml_installed = False
    print("PyYAML is not installed. Please run: pip install pyyaml")

import os
from pathlib import Path

# Check if all required packages are installed
if not all([st_installed, stauth_installed, yaml_installed]):
    print("Please install all required packages before running this application.")
    print("Run: pip install -r requirements.txt")
    exit(1)

# Set page config
st.set_page_config(
    page_title="AI Elderly Medicare System",
    page_icon="🏥",
    layout="wide",
    initial_sidebar_state="expanded"
)

# Load config file
config_path = Path(__file__).parent / "config.toml"
if config_path.exists():
    with open(config_path) as file:
        config = yaml.load(file, Loader=SafeLoader)
else:
    # Default config if file doesn't exist
    config = {
        'credentials': {
            'usernames': {
                'admin': {
                    'name': 'Admin User',
                    'password': 'admin123'
                },
                'caregiver': {
                    'name': 'Caregiver User',
                    'password': 'caregiver123'
                },
                'patient': {
                    'name': 'Patient User',
                    'password': 'patient123'
                }
            }
        },
        'cookie': {
            'expiry_days': 30,
            'key': 'ai_elderly_medicare_key',
            'name': 'ai_elderly_medicare_cookie'
        },
        'preauthorized': {
            'emails': []
        }
    }

# Create authenticator
authenticator = stauth.Authenticate(
    config['credentials'],
    config['cookie']['name'],
    config['cookie']['key'],
    config['cookie']['expiry_days'],
    config['preauthorized']
)

# Authentication
name, authentication_status, username = authenticator.login('Login', 'main')

if authentication_status:
    # Authenticated user
    st.sidebar.title(f"Welcome {name}")
    authenticator.logout('Logout', 'sidebar')
    
    # Main app content
    st.title("🏥 AI Elderly Medicare System")
    
    st.markdown("""
    ## Welcome to the AI Elderly Medicare System
    
    This system provides comprehensive healthcare management for elderly patients using AI-powered tools and services.
    
    ### Features:
    - 🏠 **Dashboard**: Overview of patient health metrics and system status
    - 👤 **Patient Management**: Patient records, medical history, and care plans
    - 💊 **Medications**: Medication tracking, reminders, and adherence monitoring
    - 📅 **Appointments**: Scheduling and teleconsultations
    - 🔔 **Notifications**: Alerts and reminders for patients and caregivers
    - 🚚 **Delivery Tracking**: Medication and supply delivery monitoring
    - 📊 **Analytics**: Health data analysis and reporting
    - ⚙️ **Settings**: User preferences and system configuration
    
    Please use the sidebar to navigate through the different sections of the application.
    """)
    
    # Display user role-based information
    if username == 'admin':
        st.info("You are logged in as an Administrator with full system access.")
    elif username == 'caregiver':
        st.info("You are logged in as a Caregiver with patient management access.")
    elif username == 'patient':
        st.info("You are logged in as a Patient with personal health access.")
        
elif authentication_status == False:
    st.error('Username/password is incorrect')
elif authentication_status == None:
    st.warning('Please enter your username and password')

# Footer
st.markdown("---")
st.markdown("💻 AI Elderly Medicare System | 🏥 Improving Elderly Healthcare with AI")

"""
AI Elderly Medicare System - Main Streamlit App
"""

def main():
    # Since we're having import issues, we'll create a simple placeholder
    # In a real implementation, this would be a full Streamlit app
    
    print("AI Elderly Medicare System - Streamlit App")
    print("==========================================")
    print()
    print("To run this application, please install the required packages:")
    print("pip install -r requirements.txt")
    print()
    print("Then run with:")
    print("streamlit run app.py")
    print()
    print("Application Structure:")
    print("- Dashboard")
    print("- Patient Management")
    print("- Medications")
    print("- Appointments")
    print("- Notifications")
    print("- Delivery Tracking")
    print("- Analytics")
    print("- Settings")

if __name__ == "__main__":
    main()
