ai-elderly-medicare-system/streamlit_app/requirements.txt
streamlit==1.22.0
streamlit-authenticator==0.2.1
pyyaml==6.0
pandas==1.5.3
numpy==1.24.3
plotly==5.14.1